+++ packages/image/src/style-native.js
@@ -1,7 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-  imageBackground: { width: "100%", height: "100%" }
-});
-
-export default styles;