+++ packages/jest-serializer/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.0"></a>
+# [0.1.0](https://github.com/newsuk/times-components/compare/@times-components/jest-serializer@0.0.2...@times-components/jest-serializer@0.1.0) (2018-04-23)
+
+
+### Features
+
+* paidOnly videos and new video package ([#889](https://github.com/newsuk/times-components/issues/889)) ([2e57f4b](https://github.com/newsuk/times-components/commit/2e57f4b))
+
+
+
+
 <a name="0.0.2"></a>
 ## 0.0.2 (2018-04-17)