+++ packages/storybook/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.9.0"></a>
+# [0.9.0](https://github.com/newsuk/times-components/compare/@times-components/storybook@0.8.0...@times-components/storybook@0.9.0) (2018-04-23)
+
+
+### Features
+
+* paidOnly videos and new video package ([#889](https://github.com/newsuk/times-components/issues/889)) ([2e57f4b](https://github.com/newsuk/times-components/commit/2e57f4b))
+
+
+
+
 <a name="0.8.0"></a>
 # [0.8.0](https://github.com/newsuk/times-components/compare/@times-components/storybook@0.7.8...@times-components/storybook@0.8.0) (2018-04-17)