+++ packages/styleguide/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.6.1"></a>
+## [0.6.1](https://github.com/newsuk/times-components/compare/@times-components/styleguide@0.6.0...@times-components/styleguide@0.6.1) (2018-04-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/styleguide
+
 <a name="0.6.0"></a>
 # [0.6.0](https://github.com/newsuk/times-components/compare/@times-components/styleguide@0.5.16...@times-components/styleguide@0.6.0) (2018-04-17)